mv web/subpath.go to utils/subpath.go
