+++ web/static.go
@@ -18,7 +18,7 @@ import (
 
 func (w *Web) InitStatic() {
 	if *w.App.Config().ServiceSettings.WebserverMode != "disabled" {
-		UpdateAssetsSubpathFromConfig(w.App.Config())
+		utils.UpdateAssetsSubpathFromConfig(w.App.Config())
 
 		staticDir, _ := utils.FindDir(model.CLIENT_DIR)
 		mlog.Debug(fmt.Sprintf("Using client directory at %v", staticDir))
